# packages/sdk/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MakerChecker API",
    "description": "Governance + orchestration control plane for enterprise AI agents.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key (mk_...)"
      }
    },
...
  },
  "paths": {
...
    "/api/proxy/sessions": {
      "post": {
        "operationId": "openProxySession",
        "tags": [
          "proxy"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "label"
                ],
                "properties": {
                  "label": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "externalRef": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/proxy/sessions/{id}/check": {
      "post": {
        "operationId": "checkProxyAction",
        "tags": [
          "proxy"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "agentName",
                  "skillRef"
                ],
                "properties": {
                  "agentName": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "skillRef": {
                    "pattern": "^[a-z0-9][a-z0-9-]*@[1-9][0-9]*$",
                    "type": "string"
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": {}
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/proxy/sessions/{id}/record": {
      "post": {
        "operationId": "recordProxyResult",
        "tags": [
          "proxy"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "checkId"
                ],
                "properties": {
                  "checkId": {
                    "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                    "type": "string"
                  },
                  "output": {},
                  "error": {}
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/proxy/sessions/{id}/close": {
      "post": {
        "operationId": "closeProxySession",
        "tags": [
          "proxy"
        ],
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/proxy/sessions/{id}": {
      "get": {
        "operationId": "getProxySession",
        "tags": [
          "proxy"
        ],
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}